Enable KGDB for Android using the msm72k_udc USB driver
Added new I/O functions to the gadget drivers to support KGDB.
The usb function f_acm.c and u_serial.c should be able to
communicate with gdb on the host PC. The changes support
concurrent use of the adb interface.

A new kgdb parameter "kgdbretry=<delay>" is added to kgdboc.c
to allow kgdb to retry the USB device specified in kgdboc when
it is initialized later in system boot. The <delay> value is
in seconds.